+++ mobile/android/android-components/docs/changelog.md
@@ -51,6 +51,8 @@ permalink: /changelog/
 * **browser-toolbar**
   * Added option to configure fading edge length by using `browserToolbarFadingEdgeSize` XML attribute.
   * Added `BrowserToolbar` attribute `browserToolbarCancelColor` to color the cancel icon.
+  * `BrowserToolbar` `cancelView` is now `clearView` with new text clearing behavior and color attribute updated from `browserToolbarCancelColor` to `browserToolbarClearColor`
+
 
 * **feature-toolbar**
   * `ToolbarPresenter` now handles situations where no `Session` is selected.